langchain-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.16.0
openai==1.51.2
pydantic==2.9.2
python-dotenv==1.0.1

